Add port reservation rule and improve PortRegistry (#8376)
* Add port reservation rule and improve PortRegistry

* Better handling for invalid calls to getPort()
