+++ tests/src/test/java/alluxio/testutils/IntegrationTestUtils.java
@@ -134,12 +134,12 @@ public final class IntegrationTestUtils {
     for (ServiceType service : Arrays.asList(ServiceType.MASTER_RPC, ServiceType.MASTER_WEB,
         ServiceType.JOB_MASTER_RPC, ServiceType.JOB_MASTER_WEB)) {
       PropertyKey key = service.getPortKey();
-      ServerConfiguration.set(key, PortRegistry.INSTANCE.reservePort());
+      ServerConfiguration.set(key, PortRegistry.reservePort());
     }
   }
 
   public static void releaseMasterPorts() {
-    PortRegistry.INSTANCE.clear();
+    PortRegistry.clear();
   }
 
   private IntegrationTestUtils() {} // This is a utils class not intended for instantiation
